Admin Interface - Kurtosis Display, Better CSS, Statements, Stylecss
Style.css update to include fontsawesome + CDNs
Kurtosis is now properly displayed on edit page
Improved UI slightly
Added statements to schema (backend)

+++ v2/src/app/components/index/index.component.ts
@@ -24,10 +24,12 @@ export class IndexComponent implements OnInit {
   constructor(private surveyservice: SurveyService, private router: Router) { }
 
   deleteSurvey(id) {
-    this.surveyservice.deleteSurvey(id).subscribe(res => {
-        this.ngOnInit();
-        console.log('Deleted');
-    });
+    if (window.confirm('Are you sure you wish to delete this survey?')){
+      this.surveyservice.deleteSurvey(id).subscribe(res => {
+          this.ngOnInit();
+          console.log('Deleted');
+      });
+    }
   }
 
   ngOnInit() {
